requirements.txt
# Base
alembic==0.9.3
Flask==0.12.2
Flask-Migrate==2.0.4
Flask-Script
Flask-SQLAlchemy==2.2
flask-restful==0.3.6
grequests
gunicorn==19.7.1
mysqlclient==1.3.10
requests
Werkzeug==0.12.2
PyJWT
./ibapi-9.76.1.tar.gz

# Dev
flake8
pytest
Jinja2
apidoc